# replace bytes.Compare with faster comparison for tag.