Record.Tag() is showing up in my profiling. This drastically reduces the time spent there:

The time for the code below gives:
```
benchmark            old ns/op     new ns/op     delta
BenchmarkStats-4     28.0          7.36          -73.71%

benchmark            old allocs     new allocs     delta
BenchmarkStats-4     0              0              +0.00%

benchmark            old bytes     new bytes     delta
BenchmarkStats-4     0             0             +0.00%
```


```Go
package main

import (
    "log"
    "testing"

    "github.com/biogo/hts/sam"
)

func main() {
}

func mustAux(a sam.Aux, err error) sam.Aux {
    if err != nil {
        panic(err)
    }
    return a
}

func BenchmarkStats(b *testing.B) {
    r := &sam.Record{
        AuxFields: []sam.Aux{
            mustAux(sam.NewAux(sam.NewTag("NM"), "AAAAAAAAAAAAAAAA")),
            mustAux(sam.NewAux(sam.NewTag("XA"), "AAAAAAAAAAAAAAAA")),
            mustAux(sam.NewAux(sam.NewTag("MC"), "AAAAAAAAAAAAAAAA")),
            mustAux(sam.NewAux(sam.NewTag("SA"), "ref,29,-,6H5M,17,0;")),
        },
    }
    key := []byte("SA")
    b.ReportAllocs()

    for i := 0; i < b.N; i++ {
        t, ok := r.Tag(key)
        if !ok {
            log.Fatal("didn't find tag")
        }
        _ = t

    }
}
```